# openapi.yaml
openapi: 3.0.0
paths:
...
  '/api/admin/states/${id}':
    delete:
      tags:
        - ADMIN
      summary: 'Удалить состояние проекта'
      operationId: b31a515315f2c36013e9295c8f24c939
      parameters:
        -
          name: id
          in: path
          description: 'ID состояния проекта'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Состояние проекта удалено'
        '404':
          description: 'Состояние проекта не найдено'
    patch:
      tags:
        - ADMIN
      summary: 'Обновить состояние проекта'
      operationId: 128656e872757229a6c40b8b7cd0f630
      parameters:
        -
          name: id
          in: path
          description: 'ID состояния проекта'
          required: true
          schema:
            type: integer
      requestBody:
        description: 'Параметры для обновления состояния проекта'
        required: true
        content:
          application/json:
            schema:
              properties:
                state:
                  type: string
              type: object
      responses:
        '200':
          description: 'Состояние проекта изменено'
        '404':
          description: 'Состояние проекта не найдено'
...
  '/api/admin/supervisors/${id}':
    delete:
      tags:
        - ADMIN
      summary: 'Удалить преподавателя'
      operationId: 80394ebee6acccdced5d5b9557058252
      parameters:
        -
          name: id
          in: path
          description: 'ID преподавателя'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Преподаватель удален'
        '404':
          description: 'Преподаватель не найден'
...
  '/api/admin/types/${id}':
    delete:
      tags:
        - ADMIN
      summary: 'Удалить тип проекта'
      operationId: ac4c955ae2fa550c7e3df8bfaea48f18
      parameters:
        -
          name: id
          in: path
          description: 'ID типа проекта'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Тип проекта удален'
        '404':
          description: 'Тип проекта не найден'
    patch:
      tags:
        - ADMIN
      summary: 'Обновить тип проекта'
      operationId: ddc84a394082d962b2d340fb2913c855
      parameters:
        -
          name: id
          in: path
          description: 'ID типа проекта'
          required: true
          schema:
            type: integer
      requestBody:
        description: 'Параметры для создания типа проекта'
        required: true
        content:
          application/json:
            schema:
              properties:
                type:
                  type: string
              type: object
      responses:
        '200':
          description: 'Тип проекта изменен'
        '404':
          description: 'Тип проекта не найден'
  /api/admin/types:
    post:
      tags:
        - ADMIN
      summary: 'Создать тип проекта'
      operationId: c939d2a21adb706922ffabca2284bbfa
      requestBody:
        description: 'Параметры для создания типа проекта'
        required: true
        content:
          application/json:
            schema:
              properties:
                type:
                  type: string
              type: object
      responses:
        '200':
          description: 'Тип проекта создан'
...
  '/api/participations/${id}':
    post:
      tags:
        - UNUSED
      summary: 'Создание заявки на проект'
      operationId: 301a772e316b0ba3f6c6f0c6f5916c7b
      parameters:
        -
          name: id
          in: path
          description: 'ID проекта'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Заявка создана'
        '404':
          description: 'Не найдено'
        '403':
          description: 'Вы уже подали 3 заявки'
    delete:
      tags:
        - Participation
      summary: 'Удаление заявки у авторизованного студента'
      operationId: 6700919707cec7e6144a07d3c23a342e
      parameters:
        -
          name: id
          in: path
          description: 'ID Заявки'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Заявка удалена'
        '403':
          description: 'Вы не можете удалить чужую заявку'
    patch:
      tags:
        - Participation
      summary: 'Изменение приоритета заявки у авторизованного студента'
      operationId: f8565721843057e358bbc347e64c2eda
      parameters:
        -
          name: id
          in: path
          description: 'ID Заявки'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                priority:
                  type: integer
              type: object
      responses:
        '200':
          description: 'Заявка изменена'
        '403':
          description: 'Вы не можете изменить чужую заявку'
...
  '/api/candidates/{id}':
    get:
      tags:
        - Candidate
      summary: 'Получить информацию о студенте по ID'
      operationId: 24960ff0fb77086529a54ad57a10816c
      parameters:
        -
          name: id
          in: path
          description: 'ID студента'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Информация о студенте'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '404':
          description: 'Информация о студенте не найдена'
...
components:
  schemas:
    Candidate:
      properties:
        id:
          description: 'ID студента'
          type: integer
        fio:
          description: 'ФИО студента'
          type: string
        about:
          description: 'Информация о студенте'
          type: string
        email:
          description: Почта
          type: string
        numz:
          description: 'Номер зачетной книжки'
          type: string
        phone:
          description: Телефон
          type: string
        course:
          description: Курс
          type: integer
        training_group:
          description: 'Учебная группа'
          type: string
        canSendParticipations:
          description: 'Может ли отправлять заявки на проекты'
          type: boolean
      type: object
...